Regenerate the migration, with unicode literals
Regenerate the 0001_initial migration, but this time with
unicode_literals imports in the models. Since Django always uses unicode
literals in the migration anyway, this avoids embedding byte strings
(b'') in the migration code, which makes makemigrations not be noisy
under Python3.

This change should be no-op for the database, and hence regenerating it
rather than generating a new one.

This concludes all changes required for migrations, and thus fixes #16.