Switch from pre-Django 1.8 optparse, to argparse
Per Django docs, defining option_list with a list of optparse's
make_option()s was deprecated in favor of defining an add_arguments()
method that accepts an (argparse-compatible) parser argument.

While at it, also stop passing interactive=False from tests; this option
doesn't exist.

This should fix #18. The bug calls for tests, but we already have
test_sync, and newer versions of Django (>= 2.0) barf on unknown
arguments, so that implicitly fixes that.