Add ManyToManyFields on concrete models
In the concrete models, add ManyToManyFields between Network and
Facility/IXLan, and InternetExchange and Facility.

Intermediate models already exist (NetworkFacility, NetworkIXLan and
InternetExchangeFacility respectively) and these are being used as
"through" models with Django's many-to-many field.

Effectively, this does not really change the schema, but provides
forward and reverse managers to go from one of these objects to the
next. So for example, one can use ixp.fac_set.all() to list all
facilities for an exchange, or fac.ix_set.all() to do the inverse.

Note that this change really makes the definitions of the
InternetExchangeFacility and InternetExchangeFacilityBase models moot,
as they're "through" models with no additional fields, and if they were
absent, Django would create the intermediate tables in the database
itself. However, avoid not being removed in this change to minimize the
amount of changes and risk involved here.

Finally, although a more descriptive "facilities" or "networks" would be
preferrable in the author's opinion, the new fields are named
"<shorthand>_set" (fac_set, ix_set, ixlan_set) to match existing
conventions.