Support 2021-07 (#70)
* Add __str__ methods to more models

This adds an __str__ method to the following models:
 - ContactBase
 - InternetExchangeFacility
 - IXLan
 - IXLanPrefix
 - NetworkContact
 - NetworkFacility
 - NetworkIXLan

This is useful for debugging e.g. in the Django shell or with Django's
admin interface. Instead of "ContactBase object NNNN", one sees a proper
name, which can be especially helpful when a list of objects is listed.

* Add a Django admin interface view

This is rudimentary, but already super helpful in exploring the local
data and debugging all kinds of issues.

* only set up admin views if concrete models are loaded

* add py39 and django 3.2 to tests

* pyupgrade

* black format

* flake8

* isort and black

* need __version__ imported after all

* linting fixes

* changelog

* remove double import

* update deps

* rm test requirements

* to poetry

* use tox for gh actions

Co-authored-by: Faidon Liambotis <paravoid@debian.org>
Co-authored-by: Matt Griswold <grizz@20c.com>